backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
requests
PyPDF2
google-generativeai
tavily-python
python-dotenv
clerk-sdk-python
pgvector
python-multipart